Initial commit for solidity, verison 0.8.11
Summary:
Resolves T9709

Depends on D12512

This version (and possibily many versions after this) uses a custom-built cvc4 (version 1.8) because cvc4 will not be included in the official Solus repository (as discussed in T9708). cvc4 is an optional but recommended dependency that is used with SMT checker. cvc4 is first built and installed in the "setup" part, and after that the solidity package will build. Because of this, the cvc4 executable is included with this package.

Some optimization flags and dependency (`pkgconfig(cln)`) are appended to the build of cvc4 (line 29) after a member of cvc5 kindly provided some suggestions. See [cvc5/cvc5#6836](https://github.com/cvc5/cvc5/issues/6836) for details.

Test Plan:
Successfully built the [example projects](https://docs.soliditylang.org/en/latest/solidity-by-example.html) on the official Solidity documentation page.

Signed-off-by: Gavin Zhao <gavinzhaojw@protonmail.com>

Reviewers: Harvey

Reviewed By: Harvey

Subscribers: Harvey, JoshStrobl, DataDrake

Maniphest Tasks: T9709

Differential Revision: https://dev.getsol.us/D11416